# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AstroHealth API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /logs:
    get:
      operationId: listLogs
      tags: [logs]
      summary: List health log entries
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
        - name: offset
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HealthLog"
    post:
      operationId: createLog
      tags: [logs]
      summary: Create a health log entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLogBody"
      responses:
        "201":
          description: Created log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthLog"
  /logs/{id}:
    get:
      operationId: getLog
      tags: [logs]
      summary: Get a health log entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Health log entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthLog"
    patch:
      operationId: updateLog
      tags: [logs]
      summary: Update a health log entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLogBody"
      responses:
        "200":
          description: Updated log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthLog"
    delete:
      operationId: deleteLog
      tags: [logs]
      summary: Delete a health log entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Insights / aggregates ────────────────────────────────────────────────────
...
  /natal-chart:
    get:
      operationId: getNatalChart
      tags: [natal]
      summary: Get the saved natal chart with computed positions
      responses:
        "200":
          description: Natal chart data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NatalChartData"
        "404":
          description: No natal chart saved yet
    post:
      operationId: upsertNatalChart
      tags: [natal]
      summary: Create or update natal chart birth data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertNatalChartBody"
      responses:
        "200":
          description: Created or updated natal chart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NatalChartData"
...
  /natal-chart/transits:
    get:
      operationId: getNatalTransits
      tags: [natal]
      summary: Get current planetary transits aspecting natal positions
      responses:
        "200":
          description: List of transit aspects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TransitAspect"
        "404":
          description: No natal chart saved yet
...
components:
  schemas:
...
    HealthLog:
      type: object
      properties:
        id:
          type: integer
        loggedAt:
          type: string
          format: date-time
        type:
          type: string
          enum: [supplement, activity, symptom, mood, note]
        supplementId:
          type: ["integer", "null"]
        activityId:
          type: ["integer", "null"]
        supplementName:
          type: ["string", "null"]
        activityName:
          type: ["string", "null"]
        dosageTaken:
          type: ["string", "null"]
        durationMinutes:
          type: ["integer", "null"]
        intensity:
          type: ["integer", "null"]
        mood:
          type: ["integer", "null"]
        energyLevel:
          type: ["integer", "null"]
        symptoms:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        astroSnapshot:
          type: ["string", "null"]
        transcribedFrom:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required: [id, loggedAt, type, createdAt]
    CreateLogBody:
      type: object
      properties:
        loggedAt:
          type: ["string", "null"]
          format: date-time
        type:
          type: string
          enum: [supplement, activity, symptom, mood, note]
        supplementId:
          type: ["integer", "null"]
        activityId:
          type: ["integer", "null"]
        supplementName:
          type: ["string", "null"]
        activityName:
          type: ["string", "null"]
        dosageTaken:
          type: ["string", "null"]
        durationMinutes:
          type: ["integer", "null"]
        intensity:
          type: ["integer", "null"]
        mood:
          type: ["integer", "null"]
        energyLevel:
          type: ["integer", "null"]
        symptoms:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        transcribedFrom:
          type: ["string", "null"]
      required: [type]
    UpdateLogBody:
      type: object
      properties:
        loggedAt:
          type: ["string", "null"]
          format: date-time
        type:
          type: string
          enum: [supplement, activity, symptom, mood, note]
        supplementId:
          type: ["integer", "null"]
        activityId:
          type: ["integer", "null"]
        supplementName:
          type: ["string", "null"]
        activityName:
          type: ["string", "null"]
        dosageTaken:
          type: ["string", "null"]
        durationMinutes:
          type: ["integer", "null"]
        intensity:
          type: ["integer", "null"]
        mood:
          type: ["integer", "null"]
        energyLevel:
          type: ["integer", "null"]
        symptoms:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]

    # ── Insights ────────────────────────────────────────────────────────────────
...
    UpsertNatalChartBody:
      type: object
      properties:
        birthDate:
          type: string
          description: Birth date (YYYY-MM-DD)
        birthTime:
          type: string
          description: Birth time in local time (HH:MM)
          default: "12:00"
        birthPlace:
          type: string
        birthLat:
          type: number
        birthLon:
          type: number
        utcOffset:
          type: number
          description: UTC offset in hours (e.g. -5 for EST, +1 for CET)
          default: 0
      required: [birthDate, birthPlace, birthLat, birthLon]
    NatalChartData:
      type: object
      properties:
        id:
          type: integer
        birthDate:
          type: string
        birthTime:
          type: string
        birthPlace:
          type: string
        birthLat:
          type: number
        birthLon:
          type: number
        utcOffset:
          type: number
        ascendant:
          $ref: "#/components/schemas/NatalPoint"
        midheaven:
          $ref: "#/components/schemas/NatalPoint"
        planets:
          type: array
          items:
            $ref: "#/components/schemas/NatalPlanetData"
        houses:
          type: array
          items:
            $ref: "#/components/schemas/HouseData"
        createdAt:
          type: string
          format: date-time
      required: [id, birthDate, birthTime, birthPlace, birthLat, birthLon, utcOffset, ascendant, midheaven, planets, houses, createdAt]
    NatalPoint:
      type: object
      properties:
        sign:
          type: string
        degree:
          type: number
        longitude:
          type: number
      required: [sign, degree, longitude]
    NatalPlanetData:
      type: object
      properties:
        planet:
          type: string
        sign:
          type: string
        degree:
          type: number
        retrograde:
          type: boolean
        longitude:
          type: number
        houseNumber:
          type: integer
      required: [planet, sign, degree, retrograde, longitude, houseNumber]
    HouseData:
      type: object
      properties:
        number:
          type: integer
        sign:
          type: string
        cuspDegree:
          type: number
        planets:
          type: array
          items:
            type: string
      required: [number, sign, cuspDegree, planets]
...
    TransitAspect:
      type: object
      properties:
        transitPlanet:
          type: string
        transitSign:
          type: string
        natalPlanet:
          type: string
        natalSign:
          type: string
        natalHouse:
          type: integer
        aspect:
          type: string
        orb:
          type: number
        healthNote:
          type: string
        exact:
          type: boolean
        score:
          type: integer
          description: Composite severity score (planet weight + aspect weight + orb weight + natal sensitivity + house bonus)
        severity:
          type: string
          enum: [mild, moderate, strong, major]
        likelyDomains:
          type: array
          items:
            type: string
      required: [transitPlanet, transitSign, natalPlanet, natalSign, natalHouse, aspect, orb, healthNote, exact, score, severity, likelyDomains]

    # ── Natal Blueprint ───────────────────────────────────────────────────────────
...